+++ client/python/test/test_client.py
@@ -74,6 +74,14 @@ class TestClientInitialization:
         )
         assert {"authorization": "Bearer abcde"}.items() <= client.headers.items()
 
+    def test_many_endpoint_demo_loads_quickly(self, many_endpoint_demo):
+        import datetime
+
+        start = datetime.datetime.now()
+        with connect(many_endpoint_demo):
+            pass
+        assert (datetime.datetime.now() - start).seconds < 5
+
 
 class TestClientPredictions:
     @pytest.mark.flaky
@@ -1083,8 +1091,8 @@ class TestAPIInfo:
                             {
                                 "label": "name",
                                 "parameter_name": "name",
-                                "parameter_has_default": True,
-                                "parameter_default": "",
+                                "parameter_has_default": False,
+                                "parameter_default": None,
                                 "type": {"type": "string"},
                                 "python_type": {"type": "str", "description": ""},
                                 "component": "Textbox",
@@ -1117,8 +1125,8 @@ class TestAPIInfo:
                             {
                                 "label": "name",
                                 "parameter_name": "name",
-                                "parameter_has_default": True,
-                                "parameter_default": "",
+                                "parameter_has_default": False,
+                                "parameter_default": None,
                                 "type": {"type": "string"},
                                 "python_type": {"type": "str", "description": ""},
                                 "component": "Textbox",