+++ gradio/components/bar_plot.py
@@ -2,12 +2,14 @@
 
 from __future__ import annotations
 
-from typing import Any, Callable, Literal
+from typing import TYPE_CHECKING, Any, Callable, Literal
 
-import pandas as pd
 from gradio_client.documentation import document
 
 from gradio.components.plot import AltairPlot, AltairPlotData, Plot
+
+if TYPE_CHECKING:
+    import pandas as pd
 
 
 @document()
@@ -304,4 +306,6 @@ class BarPlot(Plot):
         return None
 
     def example_value(self) -> Any:
+        import pandas as pd
+
         return pd.DataFrame({self.x: [1, 2, 3], self.y: [4, 5, 6]})